Model: Move all SQL stuff to inner SqlDatabaseModel
